Restrict in-band registration to auth'd admins.
XEP-0077 provides a specification for the in-band creation of new users, and this is usually done by an unauthenticated client (since the user who will be authenticating does not yet exist). It may be desirable in a client application to allow *only* authenticated administrators (or bots with administrative privileges) to create new accounts on behalf of other users.

This change is very similar to what is described by badlop here:
	http://www.ejabberd.im/node/5020
Note that it was also necessary to modify mod_register's ip_to_string, which was used when sending registration notifications to the configured accounts, because the source for these authenticated registrations is the admin's JID and not the client's IP.

Also note that the registration examples for SleekXMPP needed to be modified to account for this change (although not the plugin implementation itself), and similar changes may be necessary for other client libraries.

The following links may also be useful context:
	http://www.ejabberd.im/node/1211
	http://yo.jabber.ru/bugzilla/show_bug.cgi?id=300
	http://stackoverflow.com/questions/7709658/asmack-xmpp-new-user-registration/7725400