Components w/out session IPs cann use mod_register
I think this is what was happening: the try_register function was using check_ip_access to determine if the registration request was indeed coming from the same IP as that stored in mnesia for the session of the registrar user, and it was failing for the registrar component because ejabberd was not storing a session IP for those connections.

I think this should safely solve it: the try_register function now checks an ip_check_exempt access control list that should be specified for mod_register in ejabberd.cfg, and it will permit registration requests made by users/components on that list even if their IP does not have a corresponding session.

Now, if the JID that is attempting to be registered isn't valid, or if the registrar both has an invalid IP and is not on that ACL, then the registration will be forbidden. Otherwise the JID is valid, and we know that one of the other two conditions was successful, so we can continue. Note this is a little hacky, since we're assuming that 'deny' is the value for failure and that all other values indicate success for the latter two checks, but I couldn't figure out syntax for a more cleanly-structured case statement that didn't duplicate the contents of the expression that continues with the registration.

Note: These changes to ejabberd.cfg.example also contain the changes from a previous commit (d58f21a404aef6394a193ee8c3320a08eb2f6e1b) to allow for amdin users to register other users.