Let clients send <iq type='result'> to ejabberd
This isn't strictly necessary, but should cut down on ejabberd's workload when sending roster pushes to the clients, which is currently a performance bottleneck in my application. Before, ejabberd was refusing to handle stanzas (spec'd here: http://xmpp.org/rfcs/rfc6121.html#roster-syntax-actions-push) that it didn't need to refuse, and sending error messages back to the client.