Add module to signal a JID when users come online
This adds a simple and somewhat-inelegant module that will send a signal message to the specified JID when any user opens a new session. Both the user and signal message can be specified, and the username (but not the full JID) of the online user is appended to the signal. I'm sure there are more XMPP-like ways to do this, but for now it doesn't matter.

This should probably be in another repo somewhere, but it doesn't seem worth the hassle for just one file.

Right now I'm getting multiple presence available stanzas whenever a user comes online, so I don't have any way to know which should trigger the appropriate actions. By sending a single special signal, I'll be able to speed things up considerably.